[CatalogPromotion] Move listener from promo bundle to core
