+++ src/Sylius/Bundle/PromotionBundle/Resources/config/services.xml
@@ -83,7 +83,7 @@
             <tag name="validator.constraint_validator" alias="sylius_coupon_generation_amount_validator" />
         </service>
 
-        <service id="Sylius\Bundle\PromotionBundle\Subscriber\CatalogPromotionSubscriber">
+        <service id="CatalogPromotionEventListener">
             <argument type="service" id="sylius.event_bus" />
             <tag name="doctrine.event_subscriber" connection="default" />
         </service>
